Make JS use 'relative' references
This is done in preparation for split reports. Make the JS works on
relative element, ie. specifying the table, rows, node that we are
working on. Indeed, once reports will be split, there will be more than
one eg. result table in the HTML, so we need to pass along on which one
we are working.