Allow opening generated html report in browser
By default terminals would open files in editors and URLs in browsers.

By printing the generated html report using an URL (file://) we enable
users to click the terminal to open it. Previously this opened the file
in the text editor which is very unlikely to be what user would want.