fix: set utf-8 encoding
Cause utf-8 is default at python3, the texts transfered among python are
conflict in encoding on Windows with Chinese(cp936, GBK), which is not
use utf-8 as default. This may cause a fact that html files write in
GBK, accoding to io module chooice, will be read as utf-8 in browsers,
which cause grabled signal.

Signed-off-by: Next Alone <12210746+NextAlone@users.noreply.github.com>